refactor: split aggregate query result helpers
